learnt basic

+++ src/request_input.rs
@@ -4,5 +4,5 @@ use serde::{Serialize, Deserialize};
 pub struct UserInput{
     pub email : String,
     pub password : String,
-    pub phone_no : u32
+    pub phone_no : u64
 }
